Moved test folder to root directory. Added Makefile to specify the test configurations for Mocha. Added local version of Mocha to linz.
