+++ Makefile
@@ -1,0 +1,6 @@
+test:
+	@./node_modules/.bin/mocha
+test-formtools:
+	@./node_modules/.bin/mocha \
+		test/formtools.js
+.PHONY: test
