--- lib/formtools/test/mocha.opts
@@ -1,3 +1,0 @@
---require test/test-helper.js
---timeout 1000
---reporter spec
