+++ package.json
@@ -32,6 +32,7 @@
     "chalk": "~0.4.0"
   },
   "devDependencies": {
-    "should": "3.3.1"
+    "should": "3.3.1",
+    "mocha": "1.18.2"
   }
 }
